pyproject.toml
[project]
name = "narrative-forge"
version = "0.1.0"
description = "Reimagining timeless stories across infinite worlds"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "langchain>=1.0.5",
  "langchain-google-genai>=3.0.2",
  "langgraph>=1.0.3",
  "pydantic>=2.12.4",
  "python-dotenv>=1.2.1",
  "rich>=14.2.0",
  "tabulate>=0.9.0",
]
